feat: client error interceptor
